requirements/local.txt
asgiref==3.4.1
Django==4.0.1
djangorestframework==3.12.4
django-model-utils==4.2.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2021.5
